FEAT: Completed validating stale set recipe
